pyproject.toml
[project]
name = "audioscrambler"
version = "1.0.0"
description = "Audio scramble tool - inverts the frequency spectrum (0-5kHz)"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "fpdf2>=2.8.4",
    "jupyter>=1.1.1",
    "local-ffmpeg>=0.1.4",
    "matplotlib>=3.10.6",
    "pydub>=0.25.1",
    "removeaccents>=0.1",
    "scipy>=1.16.1",
]
